+++ changelog.rst
@@ -6,7 +6,8 @@ dev
 - TextPositionSelector no longer inherits from Range
 - TextPositionSet no longer inherits from RangeSet
 - TextPositionSelector no longer has real_start and real_end that can differ from start and end
-- Seletors and TextPositionSets are Pydantic models
+- Selectors and TextPositionSets are Pydantic models
+- TextSelector is Pydantic model for either Quote or Position Selector
 
 0.5.3 (2021-08-11)
 ------------------